+++ CHANGES.rst
@@ -11,6 +11,9 @@
   compression type. (`Issue #21 <https://github.com/graingert/redshift_sqlalchemy/pull/21>`_)
 - Explicitly require SQLAlchemy >= 0.9.2 for 'dialect_options'.
   (`Issue #13 <https://github.com/graingert/redshift_sqlalchemy/pull/13>`_)
+- Refactor of UnloadFromSelect including support for specifying all documented
+  redshift options.
+  (`Issue #27 <https://github.com/graingert/redshift_sqlalchemy/pull/27>`_)
 
 
 0.1.2 (2015-08-11)